pyproject.toml
[project]
name = "directory-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp[cli]>=1.26.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
